Keep the stage packages in the corresponding part
Since the stage packages are installed at $CRAFT_PART_INSTALL
before the BUILD stage begins, we can run the clean script in
that point to remove all the duplicated files while keeping
the stage packages in their corresponding part.

This means that we have to run the cleaning script in each
part, but since we check the files in the INSTALL folder
(instead of checking all the files in the snaps and remove
whichever ones match), it is still much faster than the
previous method.